Enable Jinja do extension
set statement in Jinja does not export variables outside of it's scope,
which means that in a {% for %} loop one can't extend a list defined in an
outerscope (Funny, you that it works in an {% if %} block, must be a
bug).

A horrible workaround is to do this:

{% set d = my_list.extend([i]) %}

Which will execute the exten sideffect and assign None to d as a
throwaway variable. do is syntatic sugar for the previous, which just
ignores the return value.