Generate SSH coordinator information (incomplete) #49
Initial stab at generating SSH connection information for master user.
The goal of this solution is to store this information in a directory
within the the host and mount it inside the jupyter container.

It does not work wihtout mounting the directory within the JupyterHub
container, but in the current approach there is no simple way to augment
the list of volumes that needs to be extended.