Add a way to extend JupyterHub mountpoints #49
Besides the default mountpoints for JupyterHub that also get mounted in
Jupyter, there is the extra_user_dirs list, which provides for new
directories to be made available within the Jupyter container.

This is user by the cluster-coordinator role that is leveraging the
jupyterhub role. We are mounting the ssh credentials for the coordinator
user within the Jupyter container.